Add missing cast
This cast is already present in other places, but was missed here.

+++ hilti/toolchain/include/ast/node.h
@@ -1056,7 +1056,7 @@ private:
             end = static_cast<int>(_children.size()) + *end;
 
         if ( ! end )
-            end = _children.size();
+            end = static_cast<int>(_children.size());
 
         if ( end > begin )
             return end;
